[fix] StateEncoder and OperationEncoder
After discussion, we decided to opt for the screen-splitting approach.

The screen would show two cart at the same time. While `p1` is the cart on the left (red), `p2` is the one on the right (blue).

For state encoder module, we deleted the "SYNCING" state and "is_sync" input port.

As for operation encoder, we arrange another set of keys to perform functions for the right cart.